+++ inst/doc/FAQ.R
@@ -1,0 +1,27 @@
+## ---- eval=FALSE--------------------------------------------------------------
+#  Bacon.hist(23.45)
+#  depth23.45 <- Bacon.Age.d(23.45)
+#  hist(depth23.45)
+#  mean(depth23.45)
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  Bacon("MyCore", run=FALSE)
+#  agedepth()
+#  # or if you've set the accumulation rate prior to something different than the default:
+#  Bacon("MyCore", run=FALSE, acc.mean=c(20,5))
+#  agedepth()
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  Bacon("MyCore", thick=1)
+#  Bacon("MyCore", 1)
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  Bacon("MyCore", boundary=80, acc.mean=c(10,50))
+
+## ---- eval=FALSE--------------------------------------------------------------
+#  Bacon.hist(23.45)
+#  depth23.45 <- Bacon.Age.d(23.45)
+#  hist(depth23.45)
+#  mean(depth23.45)
+#  quantile(depth23.45, probs=c(.025, .975))
+