+++ NEWS.md
@@ -1,3 +1,9 @@
+# rbacon 2.5.7
+* added an option to agedepth to plot date labels (plot.labels)
+* added vignettes
+* If ask=FALSE, Bacon now does not ask before writing a new folder (if required)
+
+
 # rbacon 2.5.6
 * removed closeAllConnections() as requested by Kurt Hornik (CRAN)
 * adapted agedepth() for further Plum corrections